move network files into network package, and keep exe lean
