+++ src/networking.rs
@@ -1,1 +1,6 @@
-pub mod constants;
+mod client;
+mod constants;
+mod server;
+
+pub use client::echo_client;
+pub use server::echo_server;
